requirements.txt
pydantic>=2.10.6
pydantic-ai>=0.0.20 
python-dotenv
fastapi
uvicorn
pydantic-ai
logfire
opentelemetry-instrumentation-asgi
python-multipart